Refactor body components: Remove unused popups, implement new UploadPopup, ModelPopup, SaveCapturePopup, and SwitchDevicePopup. Update imports and component usage in body.tsx. Clean up styled components and related files.
