--- frontend/src/components/body/auth_popup/index.ts
@@ -1,1 +1,0 @@
-export { default as AuthPopup } from "./AuthPopup";
