+++ frontend/src/components/body/create_directory_popup/CreateDirectoryPopup.tsx
@@ -1,6 +1,13 @@
 import React, { useState } from "react";
-import { Overlay, InfoContainer, ButtonWrap } from "./indexElements";
-import { Box, CardHeader, IconButton, TextField, Button } from "@mui/material";
+import {
+  Box,
+  Dialog,
+  DialogContent,
+  IconButton,
+  TextField,
+  Button,
+  Typography,
+} from "@mui/material";
 import CloseIcon from "@mui/icons-material/Close";
 import { colours } from "@styles/colours";
 import { useBackendUrl } from "@hooks";
@@ -93,36 +100,49 @@ const CreateFolder: React.FC<params> = (props) => {
   };
 
   return (
-    <Overlay>
-      <Box
-        sx={{
-          width: "15vw",
-          height: "fit-content",
-          zIndex: 30,
-          border: `0.01vh solid LightGrey`,
-          borderRadius: 1,
-          background: colours.CFIA_Background_White,
-        }}
-        boxShadow={1}
-      >
-        <CardHeader
-          title="Create New Directory"
-          titleTypographyProps={{
-            variant: "h6",
-            align: "left",
-            fontWeight: 600,
-            fontSize: "1.3vh",
-            color: colours.CFIA_Font_Black,
-            zIndex: 30,
+    <Dialog
+      open={true}
+      onClose={handleClose}
+      maxWidth="xs"
+      fullWidth
+      slotProps={{
+        paper: {
+          sx: {
+            borderRadius: 1,
+            padding: "1vh",
+          },
+        },
+      }}
+    >
+      <DialogContent>
+        <Box
+          sx={{
+            display: "flex",
+            flexDirection: "column",
           }}
-          action={
-            <IconButton onClick={handleClose}>
+        >
+          <Box
+            sx={{
+              display: "flex",
+              justifyContent: "space-between",
+              alignItems: "center",
+              marginBottom: "2vh",
+            }}
+          >
+            <Typography
+              variant="h6"
+              sx={{
+                fontWeight: 600,
+                fontSize: "1.8vh",
+                color: colours.CFIA_Font_Black,
+              }}
+            >
+              Create New Directory
+            </Typography>
+            <IconButton onClick={handleClose} size="small">
               <CloseIcon />
             </IconButton>
-          }
-          sx={{ padding: "0.8vh 0.8vh 0.8vh 0.8vh" }}
-        />
-        <InfoContainer>
+          </Box>
           <TextField
             id="outlined-basic"
             label="Directory Name"
@@ -136,26 +156,35 @@ const CreateFolder: React.FC<params> = (props) => {
             sx={{ fontSize: "1.2vh" }}
             size="small"
           />
-          <ButtonWrap>
+          <Box
+            sx={{
+              display: "flex",
+              flexDirection: "row",
+              alignItems: "center",
+              justifyContent: "center",
+              marginTop: "2vh",
+              marginBottom: "1vh",
+              gap: "1vh",
+            }}
+          >
             <Button
               variant="outlined"
               size="medium"
               sx={{
-                marginRight: "0.9vh",
-                marginLeft: 0,
                 borderRadius: "0.4vh",
-                paddingTop: "0.3vh",
-                paddingBottom: "0.3vh",
-                paddingLeft: "0.7vh",
-                paddingRight: "0.7vh",
+                paddingTop: "0.6vh",
+                paddingBottom: "0.6vh",
+                paddingLeft: "1.5vh",
+                paddingRight: "1.5vh",
                 fontSize: "1.17vh",
                 width: "fit-content",
-                border: `0.01vh solid LightGrey`,
-                color: colours.CFIA_Font_Black,
+                border: `0.15vh solid ${colours.CFIA_Background_Blue}`,
+                color: colours.CFIA_Background_Blue,
                 "&:hover": {
-                  backgroundColor: "#F5F5F5",
-                  transition: "0.1s ease-in-out all",
-                  border: `0.01vh solid LightGrey`,
+                  backgroundColor: colours.CFIA_Background_Blue,
+                  color: colours.CFIA_Background_White,
+                  border: `0.15vh solid ${colours.CFIA_Background_Blue}`,
+                  transition: "0.2s ease-in-out all",
                 },
               }}
               onClick={() => {
@@ -168,31 +197,29 @@ const CreateFolder: React.FC<params> = (props) => {
               variant="outlined"
               size="medium"
               sx={{
-                marginRight: "0.9vh",
-                marginLeft: 0,
                 borderRadius: "0.4vh",
-                paddingTop: "0.3vh",
-                paddingBottom: "0.3vh",
-                paddingLeft: "0.7vh",
-                paddingRight: "0.7vh",
+                paddingTop: "0.6vh",
+                paddingBottom: "0.6vh",
+                paddingLeft: "1.5vh",
+                paddingRight: "1.5vh",
                 fontSize: "1.17vh",
                 width: "fit-content",
-                border: `0.01vh solid LightGrey`,
+                border: `0.15vh solid LightGrey`,
                 color: colours.CFIA_Font_Black,
                 "&:hover": {
                   backgroundColor: "#F5F5F5",
-                  transition: "0.1s ease-in-out all",
-                  border: `0.01vh solid LightGrey`,
+                  transition: "0.2s ease-in-out all",
+                  border: `0.15vh solid LightGrey`,
                 },
               }}
               onClick={handleClose}
             >
               Cancel
             </Button>
-          </ButtonWrap>
-        </InfoContainer>
-      </Box>
-    </Overlay>
+          </Box>
+        </Box>
+      </DialogContent>
+    </Dialog>
   );
 };
