+++ frontend/src/components/body/index.ts
@@ -1,11 +1,21 @@
 import BatchUploadPopup from "./batch_upload_popup/BatchUploadPopup";
 import CreateDirectoryPopup from "./create_directory_popup/CreateDirectoryPopup";
 import DeleteDirectoryPopup from "./del_directory_popup/DeleteDirectoryPopup";
-import { AuthPopup } from "./auth_popup";
+import AuthPopup from "./auth_popup/AuthPopup";
+import UploadPopup from "./load_image_popup/UploadPopup";
+import ModelPopup from "./model_popup/ModelPopup";
+import SaveCapturePopup from "./save_capture_popup/SaveCapturePopup";
+import SwitchDevicePopup from "./switch_device_popup/SwitchDevicePopup";
+import CreativeCommonsPopup from "./creative_commons_popup/CreativeCommonsPopup";
 
 export {
   BatchUploadPopup,
   CreateDirectoryPopup,
   DeleteDirectoryPopup,
   AuthPopup,
+  UploadPopup,
+  ModelPopup,
+  SaveCapturePopup,
+  SwitchDevicePopup,
+  CreativeCommonsPopup,
 };
